requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.9.0            # via ddsp, tensorboard, tensorflow, tfds-nightly
altair==4.0.1             # via streamlit
argh==0.26.2              # via watchdog
astor==0.8.1              # via streamlit, tensorflow
attrs==19.3.0             # via jsonschema, tfds-nightly
audioread==2.1.8          # via librosa
backcall==0.1.0           # via ipython
base58==2.0.0             # via streamlit
bleach==3.1.0             # via nbconvert
blinker==1.4              # via streamlit
boto3==1.11.3             # via streamlit
botocore==1.14.3          # via boto3, s3transfer, streamlit
cachetools==4.0.0         # via google-auth
certifi==2019.11.28       # via requests
cffi==1.13.2              # via soundfile
chardet==3.0.4            # via requests
click==7.0                # via streamlit
colorama==0.4.3           # via ipython
crepe==0.0.10             # via ddsp
cycler==0.10.0            # via matplotlib
ddsp==0.0.4
decorator==4.4.1          # via ipython, librosa, traitlets, validators
defusedxml==0.6.0         # via nbconvert
dill==0.3.1.1             # via tfds-nightly
docutils==0.15.2          # via botocore
entrypoints==0.3          # via altair, nbconvert
enum-compat==0.0.3        # via streamlit
future==0.18.2            # via ddsp, streamlit, tfds-nightly
gast==0.2.2               # via tensorflow
gin-config==0.3.0         # via ddsp
google-auth-oauthlib==0.4.1  # via tensorboard
google-auth==1.10.1       # via google-auth-oauthlib, tensorboard
google-pasta==0.1.8       # via tensorflow
googleapis-common-protos==1.51.0  # via tensorflow-metadata
grpcio==1.26.0            # via tensorboard, tensorflow
h5py==2.10.0              # via crepe, keras-applications
hmmlearn==0.2.3           # via crepe
idna==2.8                 # via requests
imageio==2.6.1            # via crepe
importlib-metadata==1.4.0  # via jsonschema
ipykernel==5.1.3          # via ipywidgets, notebook, pydeck
ipython-genutils==0.2.0   # via nbformat, notebook, traitlets
ipython==7.11.1           # via ipykernel, ipywidgets
ipywidgets==7.5.1         # via pydeck
jedi==0.15.2              # via ipython
jinja2==2.10.3            # via altair, nbconvert, notebook, pydeck
jmespath==0.9.4           # via boto3, botocore
joblib==0.14.1            # via librosa, scikit-learn
jsonschema==3.2.0         # via altair, nbformat
jupyter-client==5.3.4     # via ipykernel, notebook
jupyter-core==4.6.1       # via jupyter-client, nbconvert, nbformat, notebook
keras-applications==1.0.8  # via tensorflow
keras-preprocessing==1.1.0  # via tensorflow
kiwisolver==1.1.0         # via matplotlib
librosa==0.7.2            # via ddsp
llvmlite==0.31.0          # via numba
markdown==3.1.1           # via tensorboard
markupsafe==1.1.1         # via jinja2
matplotlib==3.1.2         # via crepe
mistune==0.8.4            # via nbconvert
more-itertools==8.1.0     # via zipp
nbconvert==5.6.1          # via notebook
nbformat==5.0.3           # via ipywidgets, nbconvert, notebook
notebook==6.0.2           # via widgetsnbextension
numba==0.47.0             # via librosa, resampy
numpy==1.18.1             # via altair, crepe, ddsp, h5py, hmmlearn, imageio, keras-applications, keras-preprocessing, librosa, matplotlib, numba, opt-einsum, pandas, resampy, scikit-learn, scipy, streamlit, tensorboard, tensorflow, tfds-nightly
oauthlib==3.1.0           # via requests-oauthlib
opt-einsum==3.1.0         # via tensorflow
pandas==0.25.3            # via altair, streamlit
pandocfilters==1.4.2      # via nbconvert
parso==0.5.2              # via jedi
pathtools==0.1.2          # via watchdog
pickleshare==0.7.5        # via ipython
pillow==7.0.0             # via imageio, streamlit
prometheus-client==0.7.1  # via notebook
promise==2.3              # via tfds-nightly
prompt-toolkit==3.0.2     # via ipython
protobuf==3.11.2          # via googleapis-common-protos, streamlit, tensorboard, tensorflow, tensorflow-metadata, tfds-nightly
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.19           # via cffi
pydeck==0.2.0             # via streamlit
pydub==0.23.1             # via ddsp
pygments==2.5.2           # via ipython, nbconvert
pyparsing==2.4.6          # via matplotlib
pyrsistent==0.15.7        # via jsonschema
python-dateutil==2.8.0    # via botocore, jupyter-client, matplotlib, pandas, streamlit
pytz==2019.3              # via pandas, tzlocal
pywin32==227              # via jupyter-client, jupyter-core
pywinpty==0.5.7           # via terminado
pyyaml==5.3               # via watchdog
pyzmq==18.1.1             # via jupyter-client, notebook
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.22.0          # via requests-oauthlib, streamlit, tensorboard, tfds-nightly
resampy==0.2.2            # via crepe, librosa
rsa==4.0                  # via google-auth
s3transfer==0.3.0         # via boto3
scikit-learn==0.22.1      # via crepe, hmmlearn, librosa
scipy==1.4.1              # via crepe, ddsp, hmmlearn, librosa, resampy, scikit-learn, tensorflow
send2trash==1.5.0         # via notebook
six==1.14.0               # via absl-py, bleach, cycler, ddsp, gin-config, google-auth, google-pasta, grpcio, h5py, jsonschema, keras-preprocessing, librosa, promise, protobuf, pyrsistent, python-dateutil, resampy, tensorboard, tensorflow, tfds-nightly, traitlets, validators
soundfile==0.10.3.post1   # via librosa
streamlit==0.53.0
tensorboard==2.1.0        # via tensorflow
tensorflow-estimator==2.1.0  # via tensorflow
tensorflow-metadata==0.21.0  # via tfds-nightly
tensorflow==2.1.0         # via ddsp
termcolor==1.1.0          # via tensorflow, tfds-nightly
terminado==0.8.3          # via notebook
testpath==0.4.4           # via nbconvert
tfds-nightly==1.3.2.dev202001210105  # via ddsp
toml==0.10.0              # via streamlit
toolz==0.10.0             # via altair
tornado==5.1.1            # via ipykernel, jupyter-client, notebook, streamlit, terminado
tqdm==4.41.1              # via tfds-nightly
traitlets==4.3.3          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, pydeck
tzlocal==2.0.0            # via streamlit
urllib3==1.25.7           # via botocore, requests
validators==0.14.1        # via streamlit
watchdog==0.9.0           # via streamlit
wcwidth==0.1.8            # via prompt-toolkit
webencodings==0.5.1       # via bleach
werkzeug==0.16.0          # via tensorboard
wheel==0.33.6             # via tensorboard, tensorflow
widgetsnbextension==3.5.1  # via ipywidgets
wrapt==1.11.2             # via tensorflow, tfds-nightly
zipp==1.0.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
